requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.2
bleach==4.1.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
comm==0.1.2
datasets==2.17.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.20.3
idna==3.4
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.3
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.9
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.2.1
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.26.4
overrides==7.4.0
packaging==23.1
pandas==2.2.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pip==23.3.1
pipreqs==0.4.13
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
Pygments==2.15.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
safetensors==0.4.2
Send2Trash==1.8.2
sentencepiece==0.1.99
setuptools==68.2.2
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.7.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
widgetsnbextension==4.0.5
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
